test: 로직 변경에 따른 테스트 코드 수정
