+++ app/src/test/java/woowacourse/movie/fixture/MovieFixture.kt
@@ -9,7 +9,7 @@ const val STAR_IS_BORN = "스타 이즈 본"
 fun createMovie(name: String): Movie =
     Movie(
         title = name,
-        screeningStartDate = LocalDate.of(2025, 5, 10),
-        screeningEndDate = LocalDate.of(2025, 5, 15),
+        screeningStartDate = LocalDate.of(2026, 5, 10),
+        screeningEndDate = LocalDate.of(2026, 5, 15),
         runningTime = 152,
     )
