+++ app/src/test/java/woowacourse/movie/model/SchedulerTest.kt
@@ -22,12 +22,12 @@ class SchedulerTest {
         val actual = Scheduler.screeningPeriods(movie)
         val expected =
             localDates(
-                "2025-05-10",
-                "2025-05-11",
-                "2025-05-12",
-                "2025-05-13",
-                "2025-05-14",
-                "2025-05-15",
+                "2026-05-10",
+                "2026-05-11",
+                "2026-05-12",
+                "2026-05-13",
+                "2026-05-14",
+                "2026-05-15",
             )
 
         // then
